Заключение: снимок данных подписантов и лаборатории при создании отчёта
- Метод контроля расширен до 7 значений (ВИК/РК/УК/ЦРК/ПВТ/МПК/Расслоение),
  хранится как Method + производный текст через GetMethodTitle()
- Поля подписантов (Контроль произвёл/Заключение выдал/Надзор) теперь хранят
  снимок ФИО, организации и номера удостоверения на момент выбора из списка
  сотрудников/инспекторов лаборатории, а не индекс в общий список
- nameLab/numberAttestation в Report заполняются один раз при создании
  отчёта данными Laboratory — снимок на момент выдачи заключения
- Laboratory: добавлены Master, Welder, поля лаборатории (laboratoryName,
  numberAttestation, attestationEndDate)
- ReportCreateWindow: таблицы "Заключение"/"Подписанты" в две колонки 50/50
- Шрифты: ShareTechMono объединён в общую загрузку с NotoSans, добавлена
  тема GuiStyle::Custom
- Fix: PushFont в ImGuiDatePicker под новую сигнатуру ImGui

+++ engine/source/appWindows/employeesWindow.cpp
@@ -36,7 +36,7 @@ void EmployeesWindow::Show(std::vector<Employee> &emplList)
         ImGui::TableSetupScrollFreeze(1, 1);
         ImGui::TableHeadersRow();
 
-        tableRows = emplList.size();
+        tableRows = static_cast<int>(emplList.size());
         selected.resize(tableRows);
         for (int row = 0; row < tableRows; ++row)
         {
@@ -64,7 +64,7 @@ void EmployeesWindow::Show(std::vector<Employee> &emplList)
                 rowHeight = std::max(rowHeight, ImGui::CalcTextSize(cellTexts[col]->c_str(), nullptr, false, colWidth).y);
             }
 
-            ImGui::TableSetColumnIndex(0); // возвращаемся к первому столбцу для настоящей отрисовки
+            ImGui::TableSetColumnIndex(0);                                                                 // возвращаемся к первому столбцу для настоящей отрисовки
             ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, ImGui::GetColorU32(ImGuiCol_TableHeaderBg)); /// подсвечиваем закреплённый столбец как шапку таблицы
             if (ImGui::Selectable("###", selected.at(row), ImGuiSelectableFlags_SpanAllColumns | ImGuiSelectableFlags_AllowDoubleClick, ImVec2(0, rowHeight)))
             {
